requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0